refactor: move PackageJson interface to src/types.ts
Since utils/read-pkg-up.js is part of the public API (exported via
the ./utils package export), PackageJson appears in its return type
and is effectively public. Moving it to src/types.ts makes this
explicit and consistent with the rest of the public type surface.
